+++ config/config.exs
@@ -90,7 +90,6 @@ config :td_dd, index_worker: TdDd.Search.IndexWorker
 config :td_dd, cx_index_worker: TdCx.Search.IndexWorker
 config :td_dd, dq_index_worker: TdDq.Search.IndexWorker
 config :td_dd, loader_worker: TdDd.Loader.Worker
-config :td_dd, loader_worker: TdDd.Loader.Worker
 
 config :td_dd, TdDd.Lineage, timeout: 90_000
 
@@ -110,6 +109,8 @@ config :td_cache, :event_stream,
     [group: "dd", key: "data_structure:events", consumer: TdDd.Cache.StructureLoader],
     [group: "dd", key: "template:events", consumer: TdDd.Search.IndexWorker],
     [group: "dq", key: "business_concept:events", consumer: TdDq.Search.IndexWorker],
+    [group: "dq", key: "domain:events", consumer: TdDq.Cache.DomainEventConsumer],
+    [group: "dq", key: "implementation:events", consumer: TdDq.Cache.ImplementationLoader],
     [group: "dq", key: "template:events", consumer: TdDq.Search.IndexWorker]
   ]
 
@@ -127,7 +128,6 @@ config :td_dd, TdDd.Scheduler,
              "TdDq.RuleImplementations.Migrations:cache_structures",
              "data_fields:external_ids",
              "data_structure:keys:keep",
-             "implementation:*",
              "rule_result:*",
              "source:*",
              "sources:ids_external_ids",
@@ -146,6 +146,11 @@ config :td_dd, TdDd.Scheduler,
     job_indexer: [
       schedule: "@daily",
       task: {TdCx.Search.IndexWorker, :reindex, []},
+      run_strategy: Quantum.RunStrategy.Local
+    ],
+    implementation_cache_refresher: [
+      schedule: "@hourly",
+      task: {TdDq.Cache.ImplementationLoader, :refresh, []},
       run_strategy: Quantum.RunStrategy.Local
     ],
     rule_cache_refresher: [