+++ CHANGELOG.md
@@ -5,13 +5,18 @@
 # Added
 
 - [TD-3233] Rule result remediation plan
+- [TD-4271] Support for linking implementations with business concepts
+
+## [4.40.0] 2022-03-14
 
 ### Changed
 
+- [TD-4500] Support data structure with multiple `domain_ids`
 - [TD-4461] Avoid reindexing when a domain is modified
 - [TD-4491] Refactored search and permissions
 - [TD-4604] Include actions in `POST /api/rule_implementations/search` and `POST
  /api/data_structures/search` responses
+
 
 ## [4.39.0] 2022-03-07